double numbers

+++ ex01/RPN.hpp
@@ -6,7 +6,7 @@
 /*   By: poverbec <poverbec@student.42heilbronn>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/18 10:11:50 by poverbec          #+#    #+#             */
-/*   Updated: 2025/12/18 11:44:38 by poverbec         ###   ########.fr       */
+/*   Updated: 2025/12/19 16:23:13 by poverbec         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,6 +23,6 @@
 
 bool checkValid (std::string number );
 
-int RPN(std::string number);
+float RPN(std::string number);
